Rename lang to language
Perhaps not quite idiomatic but I prefer clarity over flexing
